Extract text shaping helpers from callback.rs (refs #88)
Move shape_run, shape_and_rasterize, and byte_offset_to_col_offset
into a new shape.rs module. callback.rs drops from 1,468 → 1,185 lines.

Required visibility changes:
- CachedGlyph, TextRun → pub(crate) with pub(crate) fields
- ShapedGlyphEntry fields → pub(crate)

These types remain defined in callback.rs; shape.rs imports them.
A subsequent PR will move the type definitions out into state.rs.

No behavior changes.
